scripts/ingestion/requirements.txt
requests
beautifulsoup4
openai
supabase
tqdm
python-dotenv
pandas
tiktoken